refactor(ci-gate): use comment prompt + ready-for-aw label
GITHUB_TOKEN-added labels don't trigger other workflows (GitHub's
anti-cascade protection). Instead of auto-labeling, ci-gate now posts a
comment instructing the PR owner to manually add the 'ready-for-aw'
label to trigger agentic CI smoke tests.

Changes:
- Rename label from 'ready-for-ci' to 'ready-for-aw' across all workflows
- ci-gate posts/updates a comment instead of calling issues.addLabels
- Add 'opened' to pull_request trigger types
- Add copilot-pull-request-reviewer[bot] to reviewer Set
- Rename test file to match new label name
