+++ .github/workflows/smoke-claude.md
@@ -5,7 +5,7 @@ on:
   schedule: every 12h
   workflow_dispatch:
   label_command:
-    name: ready-for-ci
+    name: ready-for-aw
     events: [pull_request]
     remove_label: false
   reaction: "heart"
